requirements.txt
anthropic>=0.40
openai>=1.0
flask>=3.0
python-dotenv>=1.0
httpx>=0.27
